improve update per CR suggestoin + add/refactor unit tests
